Favorites add/remove

+++ app/src/main/java/org/pathfinderfr/app/database/DBHelper.java
@@ -4,11 +4,13 @@ import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.database.sqlite.SQLiteException;
 import android.database.sqlite.SQLiteOpenHelper;
 
 import org.pathfinderfr.app.database.entity.DBEntity;
 import org.pathfinderfr.app.database.entity.DBEntityFactory;
-import org.pathfinderfr.app.database.entity.SpellFactory;
+import org.pathfinderfr.app.database.entity.EntityFactories;
+import org.pathfinderfr.app.database.entity.FavoriteFactory;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -16,7 +18,6 @@ import java.util.List;
 public class DBHelper extends SQLiteOpenHelper {
 
     public static final String DATABASE_NAME = "pathfinderfr-data.db";
-    private List<DBEntityFactory> factories;
 
     private static DBHelper instance;
 
@@ -32,29 +33,29 @@ public class DBHelper extends SQLiteOpenHelper {
 
     private DBHelper(Context context) {
         super(context, DATABASE_NAME, null, 1);
-
-        factories = new ArrayList<>();
-        factories.add(SpellFactory.getInstance());
     }
 
     @Override
     public void onCreate(SQLiteDatabase db) {
-        for (DBEntityFactory f : factories) {
+        for (DBEntityFactory f : EntityFactories.FACTORIES) {
             db.execSQL(f.getQueryCreateTable());
         }
     }
 
     public void clear() {
         SQLiteDatabase db = this.getWritableDatabase();
-        for (DBEntityFactory f : factories) {
-            db.execSQL(String.format("DROP TABLE IF EXISTS %s", f.getTableName()));
+        for (DBEntityFactory f : EntityFactories.FACTORIES) {
+            // never drop favorites!
+            if(!(f instanceof FavoriteFactory)) {
+                db.execSQL(String.format("DROP TABLE IF EXISTS %s", f.getTableName()));
+            }
             db.execSQL(f.getQueryCreateTable());
         }
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        for (DBEntityFactory f : factories) {
+        for (DBEntityFactory f : EntityFactories.FACTORIES) {
             db.execSQL(String.format("DROP TABLE IF EXISTS %s", f.getTableName()));
         }
         onCreate(db);
@@ -73,6 +74,53 @@ public class DBHelper extends SQLiteOpenHelper {
         return rowId != -1;
     }
 
+    /**
+     * Inserts the favorite into the database
+     *
+     * @param entity entity instance
+     * @return true if insertion succeeded
+     */
+    public boolean insertFavorite(DBEntity entity) {
+        SQLiteDatabase db = this.getWritableDatabase();
+        DBEntityFactory factory = FavoriteFactory.getInstance();
+        ContentValues contentValues = factory.generateContentValuesFromEntity(entity);
+        long rowId = db.insert(factory.getTableName(), null, contentValues);
+        return rowId != -1;
+    }
+
+    /**
+     * Delete the favorite(s) from the database
+     *
+     * @param entity entity instance
+     * @return true if insertion succeeded
+     */
+    public boolean deleteFavorite(DBEntity entity) {
+        SQLiteDatabase db = this.getWritableDatabase();
+        String query = FavoriteFactory.getInstance().getQueryDeleteByIds(entity.getFactory().getFactoryId(), entity.getId());
+        db.execSQL(query);
+        return true;
+    }
+
+    /**
+     * @param factoryId the type of element (skill, feat, spell, etc.)
+     * @param entityId unique ID of the entity
+     * @return true if the element has been selected as favorite, false otherwise
+     */
+    public boolean isFavorite(String factoryId, long entityId) {
+        if(FavoriteFactory.FACTORY_ID.equalsIgnoreCase(factoryId)) {
+            return true;
+        }
+        SQLiteDatabase db = this.getReadableDatabase();
+        String query = FavoriteFactory.getInstance().getQueryFetchByIds(factoryId, entityId);
+        Cursor res = db.rawQuery(query, null);
+        return res.getCount() > 0;
+    }
+
+    /**
+     * @param id unique ID of the entity to be fetched
+     * @param factory factory corresponding to the element (skill, feet, spell, etc.)
+     * @return the entity as object (assuming that it will always be found)
+     */
     public DBEntity fetchEntity(long id, DBEntityFactory factory) {
         SQLiteDatabase db = this.getReadableDatabase();
         Cursor res =  db.rawQuery( factory.getQueryFetchById(id), null );
@@ -84,23 +132,33 @@ public class DBHelper extends SQLiteOpenHelper {
     public List<DBEntity> getAllEntities(DBEntityFactory factory) {
         ArrayList<DBEntity> list = new ArrayList<>();
 
-        SQLiteDatabase db = this.getReadableDatabase();
-        Cursor res = db.rawQuery(factory.getQueryFetchAll(), null);
-        res.moveToFirst();
-        while (res.isAfterLast() == false) {
-            list.add(factory.generateEntity(res));
-            res.moveToNext();
+        try {
+            SQLiteDatabase db = this.getReadableDatabase();
+            Cursor res = db.rawQuery(factory.getQueryFetchAll(), null);
+            System.out.println("Number of elements found in database: " + res.getCount());
+            res.moveToFirst();
+            while (res.isAfterLast() == false) {
+                list.add(factory.generateEntity(res));
+                res.moveToNext();
+            }
+        } catch(SQLiteException exception) {
+            exception.printStackTrace();
         }
 
         return list;
     }
 
     public long getCountEntities(DBEntityFactory factory) {
-        SQLiteDatabase db = this.getReadableDatabase();
-        String query = String.format("SELECT COUNT(*) as total FROM %s", factory.getTableName());
-        Cursor res = db.rawQuery(query, null);
-        res.moveToFirst();
-        return res.getLong(res.getColumnIndex("total"));
+        try {
+            SQLiteDatabase db = this.getReadableDatabase();
+            String query = String.format("SELECT COUNT(*) as total FROM %s", factory.getTableName());
+            Cursor res = db.rawQuery(query, null);
+            res.moveToFirst();
+            return res.getLong(res.getColumnIndex("total"));
+        } catch(SQLiteException exception) {
+            exception.printStackTrace();
+            return 0;
+        }
     }
 
 }
